Refactor : [FE] 스타일 수정 및 파일 구조 변경
